# api/openapi.runtime.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JobPilot Platform API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/jobs": {
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "List Jobs",
    "operationId": "list_jobs_api_v1_jobs_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "employment",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Employment"
      }
     },
     {
      "name": "arrangement",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Arrangement"
      }
     },
     {
      "name": "workplace",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Workplace"
      }
     },
     {
      "name": "country",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Country"
      }
     },
     {
      "name": "sponsorship",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sponsorship"
      }
     },
     {
      "name": "min_salary",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Salary"
      }
     },
     {
      "name": "posted_within_hours",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 168,
       "title": "Posted Within Hours"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "-posted_at",
       "title": "Sort"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 25,
       "title": "Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/jobs/export": {
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "Export Jobs",
    "operationId": "export_jobs_api_v1_jobs_export_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "posted_within_hours",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 48,
       "title": "Posted Within Hours"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/jobs/{job_id}": {
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "Get Job",
    "description": "Job detail: posting + heuristic/AI extraction + my match + tracking.",
    "operationId": "get_job_api_v1_jobs__job_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Job Api V1 Jobs  Job Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/search-runs": {
   "post": {
    "tags": [
     "jobs"
    ],
    "summary": "Trigger Search Run",
    "operationId": "trigger_search_run_api_v1_search_runs_post",
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchRunOut"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/search-runs/{run_id}": {
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "Get Search Run",
    "operationId": "get_search_run_api_v1_search_runs__run_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchRunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/documents": {
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "Generate Document",
    "operationId": "generate_document_api_v1_documents_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GenerateIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DocumentOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "List Documents",
    "operationId": "list_documents_api_v1_documents_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "job_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/DocumentOut"
         },
         "title": "Response List Documents Api V1 Documents Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{document_id}": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "Get Document",
    "operationId": "get_document_api_v1_documents__document_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Document Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DocumentOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/connectors": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Connectors",
    "operationId": "list_connectors_api_v1_admin_connectors_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ConnectorOut"
         },
         "type": "array",
         "title": "Response List Connectors Api V1 Admin Connectors Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/admin/connectors/{connector_id}": {
   "patch": {
    "tags": [
     "admin"
    ],
    "summary": "Update Connector",
    "operationId": "update_connector_api_v1_admin_connectors__connector_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "connector_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Connector Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ConnectorPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConnectorOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "ConnectorOut": {
    "properties": {
     "connector_id": {
      "type": "string",
      "title": "Connector Id"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "compliance_mode": {
      "type": "string",
      "title": "Compliance Mode"
     },
     "enabled": {
      "type": "boolean",
      "title": "Enabled"
     },
     "rate_limit_per_min": {
      "type": "integer",
      "title": "Rate Limit Per Min"
     }
    },
    "type": "object",
    "required": [
     "connector_id",
     "display_name",
     "compliance_mode",
     "enabled",
     "rate_limit_per_min"
    ],
    "title": "ConnectorOut"
   },
   "ConnectorPatch": {
    "properties": {
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     },
     "rate_limit_per_min": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 600.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Rate Limit Per Min"
     }
    },
    "type": "object",
    "title": "ConnectorPatch"
   },
...
   "DocumentOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "job_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Job Id"
     },
     "doc_type": {
      "type": "string",
      "title": "Doc Type"
     },
     "content_md": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Content Md"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "created_at": {
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "doc_type",
     "created_at"
    ],
    "title": "DocumentOut"
   },
...
   "GenerateIn": {
    "properties": {
     "job_id": {
      "type": "string",
      "format": "uuid",
      "title": "Job Id"
     },
     "doc_type": {
      "type": "string",
      "enum": [
       "TAILORED_RESUME",
       "COVER_LETTER",
       "RECRUITER_EMAIL",
       "LINKEDIN_MESSAGE",
       "COLD_EMAIL"
      ],
      "title": "Doc Type"
     },
     "resume_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Resume Id"
     }
    },
    "type": "object",
    "required": [
     "job_id",
     "doc_type"
    ],
    "title": "GenerateIn"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "JobOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "connector_id": {
      "type": "string",
      "title": "Connector Id"
     },
     "url": {
      "type": "string",
      "title": "Url"
     },
     "location_text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Location Text"
     },
     "workplace": {
      "type": "string",
      "title": "Workplace"
     },
     "employment": {
      "type": "string",
      "title": "Employment"
     },
     "arrangement": {
      "type": "string",
      "title": "Arrangement"
     },
     "salary_min": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Min"
     },
     "salary_max": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Max"
     },
     "salary_currency": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Currency"
     },
     "posted_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Posted At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "title",
     "connector_id",
     "url",
     "workplace",
     "employment",
     "arrangement"
    ],
    "title": "JobOut"
   },
   "JobPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/JobOut"
      },
      "type": "array",
      "title": "Items"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "size": {
      "type": "integer",
      "title": "Size"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     }
    },
    "type": "object",
    "required": [
     "items",
     "page",
     "size",
     "total"
    ],
    "title": "JobPage"
   },
...
   "SearchRunOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "stats": {
      "additionalProperties": true,
      "type": "object",
      "title": "Stats",
      "default": {}
     }
    },
    "type": "object",
    "required": [
     "id",
     "status"
    ],
    "title": "SearchRunOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}